Move `ChatTestEnv` to `...chat.given` package.
